Add usernames from syslog as commit authors
Create a syslog server to receive configuration changes syslogs from the
firewall contexts. Each syslog message contains the user that executed
the command. When syslogs are received, they are added to a per-context
buffer. If 30 seconds pass without adding a syslog to a buffer, or if a
"write memory" is received for the given context, the new context
configuration is downloaded and committed as the user who made the
changes. The commands executed are included in the commit message. As a
failsafe, the configurations are updated if they have not been updated
in 3600 seconds or more.

This also involved a rewrite of the FWSM SSH interaction to become a
"blocking" interface with respect to commands, i.e.
    config = fwsm.cmd 'show run'
    # block while config is buffering

This was accomplished by running all SSH interactions in a separate
thread and blocking the caller using Mutexes/ConditionVariables (this is
probably pretty dirty :)). After a command is executed, output lines are
buffered until a line  matching the prompt is received. Every buffered
line is considered part of the output from the command and is returned
to the caller (excluding any leading/trailing whitespace).

This implementation was created using syslogs forwarded from a syslog-ng
server. The original syslogs did not contain the hostname, but syslog-ng
adds this when the log is forwarded. This implementation would therefore
NOT work with syslogs going directly to this program.
(@TODO???)