Handle some differences in log format.
This program was developed using forwarded syslogs from a syslog-ng
server, which adds the hostname syslog-ng received the logs from to the
syslog message. (I should probably look into this).  Previously, this
had been relied on to map the commands to the originating firewall
context. It was later discovered that a configuration option exists to
populate the context name in the syslogs:

logging device-id context-name
(note: this could be configured to send device-ids other than the
context name)

This commit adds support for logs that do or do not contain the
context name. (@TODO handle pulling IP from socket?)

There is a known issue if 'logging timestamps' is also enabled;
presently the timestamps are inserted by the syslog-ng server (which
uses a different format than the timestamps :D)

The context name is currently determined as follows:

Attempt to parse the context name out of the syslog:
    - If it is not present, use the IP:context map
    - Otherwise, use the context from the syslog